* Updated docs * Updated dev-cpp files * Updated config file * Party experience share ranges and level differences are configurable * rateExperience, rateMagic and rateSkills are now floats * Added per-player experience multipliers controllable with Lua * Fixed guild commands * Added getConfigValue Lua function * Added more command line arguments, not only --config, but also --help, --port, --ip, --output-log, --error-log * Logs now work properly * Fixed NPCs interaction * Fixed spells (runes) * Fixed /online and /town talkaction-commands * MySQL read timeout is now configurable
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@1194 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ game.h
@@ -554,7 +554,7 @@ class Game
 		bool closeRuleViolation(Player* player);
 
 		bool loadExperienceStages();
-		uint64_t getExperienceStage(uint32_t level);
+		double getExperienceStage(uint32_t level);
 
 		void setGlobalSaveMessage(int16_t key, bool value) {globalSaveMessage[key] = value;}
 		bool getGlobalSaveMessage(int16_t key) const {return globalSaveMessage[key];}
@@ -622,7 +622,7 @@ class Game
 		RefreshTiles refreshTiles;
 		Trash trash;
 
-		typedef std::map<int32_t, int32_t> StageList;
+		typedef std::map<int32_t, float> StageList;
 		StageList stages;
 		uint32_t lastStageLevel;
